backend/app/scrapers/amazon_scraper.py

"""
Amazon US scraper for book data and prices.

This module implements ethical web scraping for Amazon book listings.
IMPORTANT: This is a placeholder implementation. Real scraping requires
careful consideration of robots.txt, rate limiting, and legal compliance.
"""

import aiohttp
import asyncio
from typing import List, Dict, Any, Optional
from bs4 import BeautifulSoup
import re
import random

class AmazonScraper:
    """
    Amazon scraper for book data and pricing information.
    
    NOTE: This is a placeholder implementation for development.
    Production use requires:
    1. Compliance with Amazon's robots.txt and Terms of Service
    2. Proper rate limiting and request headers
    3. Legal review and potentially Amazon API access
    4. Robust error handling and retry logic
    """
    
    def __init__(self):
        self.base_url = "https://www.amazon.com"
        self.search_url = "https://www.amazon.com/s"
        self.user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
        self.rate_limit = 1  # seconds between requests
    
    async def scrape_book(self, isbn: str) -> Optional[Dict[str, Any]]:
        """
        Scrape book data by ISBN from Amazon.
        
        Args:
            isbn: Book ISBN to search for
            
        Returns:
            Dict: Book data including price and metadata
            
        Note:
            This is a placeholder that returns mock data.
            Real implementation would parse Amazon HTML.
        """
        # Simulate API delay
        await asyncio.sleep(random.uniform(1, 3))
        
        # TODO: Replace with actual Amazon scraping
        # For now, return mock data to demonstrate the structure
        mock_data = {
            "isbn": isbn,
            "title": f"Sample Book for ISBN {isbn}",
            "author": "Sample Author",
            "publisher": "Sample Publisher",
            "publication_year": 2023,
            "category": "Technology",
            "description": f"This is a sample book description for ISBN {isbn}. In a real implementation, this would be scraped from Amazon's product page.",
            "image_url": "https://via.placeholder.com/300x400?text=Book+Cover",
            "pages": random.randint(200, 500),
            "weight": f"{random.uniform(0.5, 2.0):.1f} lbs",
            "dimensions": "8.5 x 5.5 x 1.0 inches",
            "price": round(random.uniform(10.00, 50.00), 2),
            "original_price": round(random.uniform(15.00, 60.00), 2),
            "availability": random.choice(["in_stock", "limited", "out_of_stock"]),
            "condition": "new",
            "shipping_cost": random.choice([0.00, 3.99, 5.99]),
            "url": f"https://www.amazon.com/dp/example{isbn}"
        }
        
        return mock_data
    
    async def search_books(self, query: str, max_results: int = 20) -> List[Dict[str, Any]]:
        """
        Search for books on Amazon and return results.
        
        Args:
            query: Search query
            max_results: Maximum number of results to return
            
        Returns:
            List[Dict]: List of book data from search results
            
        Note:
            This is a placeholder that returns mock data.
            Real implementation would parse Amazon search results.
        """
        # Simulate API delay
        await asyncio.sleep(random.uniform(2, 4))
        
        # TODO: Replace with actual Amazon search scraping
        # For now, return mock search results
        results = []
        for i in range(min(max_results, 10)):  # Limit mock results
            isbn = f"978{random.randint(1000000000, 9999999999)}"
            book_data = {
                "isbn": isbn,
                "title": f"{query} Related Book {i+1}",
                "author": f"Author {i+1}",
                "publisher": random.choice(["O'Reilly", "Manning", "Packt", "Apress"]),
                "publication_year": random.randint(2018, 2024),
                "category": "Technology",
                "description": f"A book related to {query}. This is mock data for development.",
                "image_url": f"https://via.placeholder.com/300x400?text=Book+{i+1}",
                "pages": random.randint(200, 500),
                "weight": f"{random.uniform(0.5, 2.0):.1f} lbs",
                "dimensions": "8.5 x 5.5 x 1.0 inches",
                "price": round(random.uniform(10.00, 50.00), 2),
                "original_price": round(random.uniform(15.00, 60.00), 2),
                "availability": random.choice(["in_stock", "limited"]),
                "condition": "new",
                "shipping_cost": random.choice([0.00, 3.99, 5.99]),
                "url": f"https://www.amazon.com/dp/example{isbn}"
            }
            results.append(book_data)
        
        return results
    
    async def _make_request(self, url: str, params: Dict[str, str] = None) -> Optional[str]:
        """
        Make HTTP request to Amazon with proper headers and rate limiting.
        
        Args:
            url: URL to request
            params: Query parameters
            
        Returns:
            str: HTML content or None if failed
            
        Note:
            This method would handle actual HTTP requests in production.
        """
        headers = {
            "User-Agent": self.user_agent,
            "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
            "Accept-Language": "en-US,en;q=0.5",
            "Accept-Encoding": "gzip, deflate",
            "Connection": "keep-alive",
        }
        
        try:
            # Rate limiting
            await asyncio.sleep(self.rate_limit)
            
            # TODO: Implement actual HTTP request
            # async with aiohttp.ClientSession() as session:
            #     async with session.get(url, headers=headers, params=params) as response:
            #         if response.status == 200:
            #             return await response.text()
            #         else:
            #             print(f"Request failed with status {response.status}")
            #             return None
            
            # For now, return None as this is a placeholder
            return None
            
        except Exception as e:
            print(f"Request error: {str(e)}")
            return None
    
    def _parse_book_page(self, html: str) -> Optional[Dict[str, Any]]:
        """
        Parse Amazon book page HTML to extract book data.
        
        Args:
            html: HTML content of book page
            
        Returns:
            Dict: Parsed book data
            
        Note:
            This method would implement actual HTML parsing in production.
        """
        # TODO: Implement HTML parsing with BeautifulSoup
        # soup = BeautifulSoup(html, 'html.parser')
        # 
        # # Extract title
        # title_elem = soup.find('span', {'id': 'productTitle'})
        # title = title_elem.text.strip() if title_elem else None
        # 
        # # Extract price
        # price_elem = soup.find('span', class_='a-price-whole')
        # price = float(price_elem.text.replace(',', '')) if price_elem else None
        # 
        # # Extract other fields...
        # 
        # return {
        #     'title': title,
        #     'price': price,
        #     # ... other fields
        # }
        
        # Placeholder return
        return None
    
    def _parse_search_results(self, html: str) -> List[Dict[str, Any]]:
        """
        Parse Amazon search results HTML to extract book listings.
        
        Args:
            html: HTML content of search results page
            
        Returns:
            List[Dict]: List of book data from search results
            
        Note:
            This method would implement actual HTML parsing in production.
        """
        # TODO: Implement search results parsing
        # soup = BeautifulSoup(html, 'html.parser')
        # results = []
        # 
        # # Find all product containers
        # products = soup.find_all('div', {'data-component-type': 's-search-result'})
        # 
        # for product in products:
        #     # Extract product data
        #     title_elem = product.find('h2', class_='a-size-mini')
        #     price_elem = product.find('span', class_='a-price-whole')
        #     
        #     if title_elem and price_elem:
        #         results.append({
        #             'title': title_elem.text.strip(),
        #             'price': float(price_elem.text.replace(',', '')),
        #             # ... other fields
        #         })
        # 
        # return results
        
        # Placeholder return
        return []

# Production scraping considerations:
"""
IMPORTANT LEGAL AND ETHICAL CONSIDERATIONS:

1. ROBOTS.TXT COMPLIANCE:
   - Always check and respect Amazon's robots.txt file
   - Follow crawl-delay directives

2. RATE LIMITING:
   - Implement proper delays between requests (1-3 seconds minimum)
   - Use exponential backoff for retry logic
   - Monitor for rate limiting responses

3. USER AGENT AND HEADERS:
   - Use realistic browser headers
   - Rotate user agents periodically
   - Include proper Accept headers

4. LEGAL COMPLIANCE:
   - Review Amazon's Terms of Service
   - Consider using Amazon's official APIs instead
   - Consult with legal team before production deployment

5. TECHNICAL CONSIDERATIONS:
   - Handle JavaScript-rendered content (consider Selenium/Playwright)
   - Implement proper error handling and logging
   - Use proxy rotation for large-scale scraping
   - Cache results to minimize requests

6. ALTERNATIVE APPROACHES:
   - Amazon Product Advertising API (official)
   - Third-party book databases (Open Library, Google Books API)
   - Partner APIs from book distributors
"""
